+++ sql/parser/select_test.go
@@ -17,7 +17,7 @@ func TestParserSelect(t *testing.T) {
 	}{
 		{"NoTable", "SELECT 1",
 			query.SelectStmt{
-				Selectors: []query.ResultField{query.ResultFieldExpr{Expr: query.Int8Value(1), ExprName: "1"}},
+				Selectors: []query.ResultField{query.ResultFieldExpr{Expr: query.IntValue(1), ExprName: "1"}},
 			}, false},
 		{"NoCond", "SELECT * FROM test",
 			query.SelectStmt{
@@ -41,27 +41,27 @@ func TestParserSelect(t *testing.T) {
 			}, false},
 		{"WithExpr", "SELECT a    > 1 FROM test",
 			query.SelectStmt{
-				Selectors: []query.ResultField{query.ResultFieldExpr{Expr: query.Gt(query.FieldSelector([]string{"a"}), query.Int8Value(1)), ExprName: "a    > 1"}},
+				Selectors: []query.ResultField{query.ResultFieldExpr{Expr: query.Gt(query.FieldSelector([]string{"a"}), query.IntValue(1)), ExprName: "a    > 1"}},
 				TableName: "test",
 			}, false},
 		{"WithCond", "SELECT * FROM test WHERE age = 10",
 			query.SelectStmt{
 				TableName: "test",
 				Selectors: []query.ResultField{query.Wildcard{}},
-				WhereExpr: query.Eq(query.FieldSelector([]string{"age"}), query.Int8Value(10)),
+				WhereExpr: query.Eq(query.FieldSelector([]string{"age"}), query.IntValue(10)),
 			}, false},
 		{"WithOrderBy", "SELECT * FROM test WHERE age = 10 ORDER BY a.b.c",
 			query.SelectStmt{
 				TableName: "test",
 				Selectors: []query.ResultField{query.Wildcard{}},
-				WhereExpr: query.Eq(query.FieldSelector([]string{"age"}), query.Int8Value(10)),
+				WhereExpr: query.Eq(query.FieldSelector([]string{"age"}), query.IntValue(10)),
 				OrderBy:   []string{"a", "b", "c"},
 			}, false},
 		{"WithOrderBy ASC", "SELECT * FROM test WHERE age = 10 ORDER BY a.b.c ASC",
 			query.SelectStmt{
 				TableName:        "test",
 				Selectors:        []query.ResultField{query.Wildcard{}},
-				WhereExpr:        query.Eq(query.FieldSelector([]string{"age"}), query.Int8Value(10)),
+				WhereExpr:        query.Eq(query.FieldSelector([]string{"age"}), query.IntValue(10)),
 				OrderBy:          []string{"a", "b", "c"},
 				OrderByDirection: scanner.ASC,
 			}, false},
@@ -69,7 +69,7 @@ func TestParserSelect(t *testing.T) {
 			query.SelectStmt{
 				TableName:        "test",
 				Selectors:        []query.ResultField{query.Wildcard{}},
-				WhereExpr:        query.Eq(query.FieldSelector([]string{"age"}), query.Int8Value(10)),
+				WhereExpr:        query.Eq(query.FieldSelector([]string{"age"}), query.IntValue(10)),
 				OrderBy:          []string{"a", "b", "c"},
 				OrderByDirection: scanner.DESC,
 			}, false},
@@ -77,23 +77,23 @@ func TestParserSelect(t *testing.T) {
 			query.SelectStmt{
 				Selectors: []query.ResultField{query.Wildcard{}},
 				TableName: "test",
-				WhereExpr: query.Eq(query.FieldSelector([]string{"age"}), query.Int8Value(10)),
-				LimitExpr: query.Int8Value(20),
+				WhereExpr: query.Eq(query.FieldSelector([]string{"age"}), query.IntValue(10)),
+				LimitExpr: query.IntValue(20),
 			}, false},
 		{"WithOffset", "SELECT * FROM test WHERE age = 10 OFFSET 20",
 			query.SelectStmt{
 				Selectors:  []query.ResultField{query.Wildcard{}},
 				TableName:  "test",
-				WhereExpr:  query.Eq(query.FieldSelector([]string{"age"}), query.Int8Value(10)),
-				OffsetExpr: query.Int8Value(20),
+				WhereExpr:  query.Eq(query.FieldSelector([]string{"age"}), query.IntValue(10)),
+				OffsetExpr: query.IntValue(20),
 			}, false},
 		{"WithLimitThenOffset", "SELECT * FROM test WHERE age = 10 LIMIT 10 OFFSET 20",
 			query.SelectStmt{
 				Selectors:  []query.ResultField{query.Wildcard{}},
 				TableName:  "test",
-				WhereExpr:  query.Eq(query.FieldSelector([]string{"age"}), query.Int8Value(10)),
-				OffsetExpr: query.Int8Value(20),
-				LimitExpr:  query.Int8Value(10),
+				WhereExpr:  query.Eq(query.FieldSelector([]string{"age"}), query.IntValue(10)),
+				OffsetExpr: query.IntValue(20),
+				LimitExpr:  query.IntValue(10),
 			}, false},
 		{"WithOffsetThenLimit", "SELECT * FROM test WHERE age = 10 OFFSET 20 LIMIT 10", nil, true},
 	}
